requirements.txt
altair
numpy
pandas
pydeck
streamlit
streamlit
mysql-connector-python
